% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{ibrrra_palette}
\alias{ibrrra_palette}
\title{Ibrrra Color Palette}
\usage{
ibrrra_palette(direction = 1)
}
\arguments{
\item{direction}{Sets the order of colors in the scale. If 1, the colors are ordered according to the default. If -1, the order of colors is reversed.}
}
\value{
A character vector of color hex values
}
\description{
This function creates a vector of n colors from a maximum of five colors. The colors are: "red", "blue", "green", "grey", and "yellow"
}
\examples{
ibrrra_palette()(1)
ibrrra_palette(direction = -1)(4)

}
