% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale.R
\name{scale_colour_ibrrra_seq_c}
\alias{scale_colour_ibrrra_seq_c}
\title{Continuous Ibrrra Color Scales for ggplot2}
\usage{
scale_colour_ibrrra_seq_c(name, direction = 1, amount = 0.2, ...)
}
\arguments{
\item{name}{name of the color}

\item{direction}{Sets the order of colors in the scale. If 1, the colors are ordered according to the default. If -1, the order of colors is reversed.}

\item{amount}{amount of lighting to be used on the color. Default is 0.2}

\item{...}{arguments to be passed to ggplot2::discrete_scale}
}
\description{
Continuous colour scale function to be used with ggplot2
}
\examples{
scale_colour_ibrrra_seq_c("red")
scale_colour_ibrrra_seq_c("red", direction = -1)
scale_colour_ibrrra_seq_c("red", amount = 0.8)
scale_colour_ibrrra_seq_c("red", direction = -1, amount = 0.8)

}
